Split `backend::piston::window` into two, separating out all gfx related code into `backend::piston::gfx`, so that `backend::piston::window` is just a simple wrapper.
Most gfx functionality now handled by `GfxContext` struct, held by the window.
