+++ examples/image.rs
@@ -6,7 +6,7 @@
 extern crate find_folder;
 
 use conrod::{widget, Colorable, Positionable, Sizeable, Widget, color};
-use conrod::backend::piston::window::{ImageSize, Flip, G2dTexture, Texture};
+use conrod::backend::piston::gfx::*;
 use conrod::backend::piston::{Window, UpdateEvent};
 use conrod::backend::piston::core_event_loop::{EventLoop, WindowEvents};
 use conrod::backend::piston::window as piston_window;
@@ -40,7 +40,7 @@ fn main() {
     // Create our `conrod::image::Map` which describes each of our widget->image mappings.
     // In our case we only have one image, however the macro may be used to list multiple.
     let image_map = image_map! {
-        (ids.rust_logo, load_rust_logo(&mut window)),
+        (ids.rust_logo, load_rust_logo(&mut window.context)),
     };
 
     // We'll instantiate the `Image` at its full size, so we'll retrieve its dimensions.
@@ -71,10 +71,10 @@ fn main() {
 }
 
 // Load the Rust logo from our assets folder.
-fn load_rust_logo(window: &mut Window) -> G2dTexture<'static> {
+fn load_rust_logo(context: &mut GfxContext) -> G2dTexture<'static> {
     let assets = find_folder::Search::ParentsThenKids(3, 3).for_folder("assets").unwrap();
     let path = assets.join("images/rust.png");
-    let factory = &mut window.factory;
-    let settings = piston_window::TextureSettings::new();
+    let factory = &mut context.factory;
+    let settings = TextureSettings::new();
     Texture::from_path(factory, &path, Flip::None, &settings).unwrap()
 }
